+++ src/Console/Command/Remove.php
@@ -2,8 +2,6 @@
 
 namespace FiiSoft\Phinx\Console\Command;
 
-use Phinx\Console\Command\AbstractCommand;
-use Phinx\Migration\MigrationInterface;
 use Phinx\Util\Util;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -44,23 +42,11 @@ EOT
     {
         $this->bootstrap($input, $output);
     
-        $environment = $input->getOption('environment');
-        if (null === $environment) {
-            $environment = $this->getConfig()->getDefaultEnvironment();
-            $output->writeln('<comment>warning</comment> no environment specified, defaulting to: ' . $environment);
-        } else {
-            $output->writeln('<info>using environment</info> ' . $environment);
-        }
+        $environment = $this->getOptionEnvironment($input, $output);
     
-        $version = $input->getArgument('target');
-        if (!$version) {
-            $output->writeln('<comment>target version is required</comment>');
-            return 2;
-        }
-    
-        if (!ctype_digit($version) || strlen($version) !== 14) { //YmdHis
-            $output->writeln('<error>target '.$version.' is not a valid version number</error>');
-            return 3;
+        $version = $this->getArgumentVersion($input, $output);
+        if (is_int($version)) {
+            return $version;
         }
     
         $env = $this->manager->getEnvironment($environment);
@@ -72,15 +58,9 @@ EOT
             $output->writeln('remove migration <info>'.$version.'</info> from database');
             
             if (isset($migrations[$version])) {
-                $now = date('Y-m-d H:i:s');
-                $env->getAdapter()->migrated($migrations[$version], MigrationInterface::DOWN, $now, $now);
+                $this->markMigrationAsUnmigrated($env, $migrations[$version]);
             } else {
-                $env->getAdapter()->execute(sprintf(
-                    'DELETE FROM %s WHERE %s = %s',
-                    $env->getSchemaTableName(),
-                    $env->getAdapter()->quoteColumnName('version'),
-                    $version
-                ));
+                $this->removeMigrationFromPhinxlog($env, $version);
             }
         }